refactor payment -> approval 도메인 이름 변경 완료
