+++ src/main/java/com/ideality/coreflow/approval/domain/aggregate/ApprovalParticipant.java
@@ -1,4 +1,4 @@
-package com.ideality.coreflow.payment.domain.aggregate;
+package com.ideality.coreflow.approval.domain.aggregate;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -9,34 +9,24 @@ import lombok.NoArgsConstructor;
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "payment")
+@Table(name = "approval_participant")
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @Builder
-public class Payment {
-
+public class ApprovalParticipant {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @Column(name = "approval_id")
+    private Long paymentId;
+
     @Column(name = "user_id")
     private Long userId;
 
-    private String title;
-
-    private Type type;
-
-    private Status status;
-
-    private String content;
+    private Role role;
 
     @Column(name = "created_at", nullable = false)
     private LocalDateTime createdAt;
-
-    @Column(name = "approved_at")
-    private LocalDateTime approvedAt;
-
-    @Column(name = "work_id")
-    private Long workId;
 }
