+++ src/main/java/com/ideality/coreflow/approval/domain/aggregate/Role.java
@@ -1,0 +1,6 @@
+package com.ideality.coreflow.approval.domain.aggregate;
+
+public enum Role {
+    APPROVER,
+    VIEWER
+}
