+++ src/main/java/com/ideality/coreflow/approval/domain/aggregate/Status.java
@@ -1,4 +1,4 @@
-package com.ideality.coreflow.payment.domain.aggregate;
+package com.ideality.coreflow.approval.domain.aggregate;
 
 public enum Status {
     PENDING,
