--- src/main/java/com/ideality/coreflow/payment/domain/aggregate/Role.java
@@ -1,6 +1,0 @@
-package com.ideality.coreflow.payment.domain.aggregate;
-
-public enum Role {
-    APPROVER,
-    VIEWER
-}
